ingestion/requirements.txt
crawl4ai>=0.4.0
pymupdf>=1.24.0
unstructured>=0.14.0
langchain-openai>=0.1.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
openai>=1.30.0
tiktoken>=0.7.0
